third_party/blink/web_tests/platform/generic/netinfo/type-change-no-listener-expected.txt

Tests that the type changes without an active listener

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS connection.type is newConnectionType
PASS connection.downlinkMax is newDownlinkMax
PASS connection.effectiveType is newEffectiveType
PASS connection.rtt is newRtt
PASS connection.downlink is newDownlinkMax
PASS successfullyParsed is true

TEST COMPLETE
